+++ src/com/amplitude/unity/plugins/AmplitudePlugin.java
@@ -63,6 +63,10 @@ public class AmplitudePlugin {
         Amplitude.getInstance().setUserProperties(ToJSONObject(jsonProperties));
     }
 
+    public static void unsetUserProperties(String propertiesList) {
+        Amplitude.getInstance().unsetUserProperties(propertiesList.split(","));
+    }
+
     public static void logRevenue(double amount) {
         Amplitude.getInstance().logRevenue(amount);
     }